refactor: move CountriesAlpha2 and LanguagesISO639 to core-types
