+++ packages/content/src/attributes/constants/index.ts
@@ -1,3 +1,3 @@
+export * from "@nmshd/core-types/src/CountriesAlpha2";
+export * from "@nmshd/core-types/src/LanguagesISO639";
 export * from "./AttributeValueTypeQueryMapping";
-export * from "./CountriesAlpha2";
-export * from "./LanguagesISO639";
