+++ packages/content/src/attributes/types/strings/AbstractCountry.ts
@@ -1,6 +1,6 @@
 import { serialize, validate } from "@js-soft/ts-serval";
+import { CountryAlpha2 } from "@nmshd/core-types/src/CountriesAlpha2";
 import { RenderHints, RenderHintsDataType, RenderHintsEditType, ValueHints, ValueHintsValue } from "../../../attributes/hints";
-import { CountryAlpha2 } from "../../constants/CountriesAlpha2";
 import { AbstractString } from "../AbstractString";
 
 export abstract class AbstractCountry extends AbstractString {
