+++ packages/core-types/src/index.ts
@@ -3,6 +3,8 @@ export * from "./CoreDate";
 export * from "./CoreError";
 export * from "./CoreId";
 export * from "./CoreIdHelper";
+export * from "./CountriesAlpha2";
+export * from "./LanguagesISO639";
 export * from "./Random";
 export * from "./references";
 export * from "./SharedPasswordProtection";
